+++ bin/full-node/src/cli.rs
@@ -47,21 +47,9 @@ use std::{net::SocketAddr, path::PathBuf};
 pub enum CliOptions {
     /// Connects to the chain and synchronizes the local database with the network.
     Run(CliOptionsRun),
-    /// Connects to an IP address and prints some information about the node.
-    NodeInfo(CliOptionsNodeInfo),
     /// Computes the 64 bits BLAKE2 hash of a string payload and prints the hexadecimal-encoded hash.
     #[structopt(name = "blake2-64bits-hash")]
     Blake264BitsHash(CliOptionsBlake264Hash),
-}
-
-#[derive(Debug, clap::StructOpt)]
-pub struct CliOptionsNodeInfo {
-    /// IP address to connect to (format: `<ip>:<port>`).
-    // Note: we accept a String rather than a SocketAddr in order to allow for DNS addresses.
-    pub address: String,
-    /// Ed25519 private key of network identity (as a seed phrase).
-    #[structopt(long, parse(try_from_str = decode_ed25519_private_key))]
-    pub libp2p_key: Option<[u8; 32]>,
 }
 
 #[derive(Debug, clap::StructOpt)]